#107: move test utils into their own dir
