+++ packages/web/src/components/app/components/full-detail/full-detail.test.jsx
@@ -2,9 +2,9 @@ import FullDetail from "./full-detail";
 import { waitFor, screen } from "@testing-library/react";
 import { sources } from "md4k-constants";
 import { vi, beforeEach } from "vitest";
-import { renderWithProviders } from "../../../../utils/render-with-providers";
-import { createMatchMedia } from "../../../../utils/create-match-media";
-import { buildThirdPartyMovieMock } from "../../../../utils/build-third-party-movie-mock";
+import { renderWithProviders } from "../../../../test-utils/render-with-providers";
+import { createMatchMedia } from "../../../../test-utils/create-match-media";
+import { buildThirdPartyMovieMock } from "../../../../test-utils/build-third-party-movie-mock";
 import { GET_THIRD_PARTY_MOVIE_FULL_DETAILS } from "../../../../graphql/queries";
 import { sourceLabels } from "../../../../constants/sources";
