+++ packages/web/src/components/app/components/list/components/list-grid/list-grid.test.jsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import ListGrid from "./list-grid";
 import { vi } from "vitest";
-import { renderWithProviders } from "../../../../../../utils/render-with-providers";
+import { renderWithProviders } from "../../../../../../test-utils/render-with-providers";
 
 vi.mock("./components/movie/movie", () => ({
   default: ({ onDeleteMovie, movie }) => (
