+++ packages/web/src/components/app/components/watched/watched.error.test.jsx
@@ -1,10 +1,10 @@
 import { Watched } from "./watched";
 import { within, screen } from "@testing-library/react";
-import { renderWithProviders } from "../../../../utils/render-with-providers";
+import { renderWithProviders } from "../../../../test-utils/render-with-providers";
 import { vi } from "vitest";
 import { GET_MOVIES } from "../../../../graphql/queries";
 import { REMOVE_MOVIE } from "../../../../graphql/mutations";
-import { buildMovieMock } from "../../../../utils/build-movie-mock";
+import { buildMovieMock } from "../../../../test-utils/build-movie-mock";
 
 vi.mock("./components/watched-movie/watched-movie", () => ({
   default: ({ movie, onDelete }) => (
